add info in validation screen
