+++ lib/l10n/intl_fr.arb
@@ -67,7 +67,7 @@
 	"enterAddress": "Entrer l'adresse",
 	"enterPublicKey": "Enter la clé publique",
 	"userCancelledOperation": "Opération annulée par l'utilisateur",
-	"transactionConfirmationFormHeader": "Envoyer la transaction ?",
+	"transactionConfirmationFormHeader": "Demande de confirmation",
 	"unknownAccount": "Le compte %1 n'existe pas.",
 	"invalidTransaction": "La transaction est invalide",
 	"invalidAddress": "L'adresse saisie n'est pas valide",
@@ -336,7 +336,7 @@
 	"selectAccountDescSeveral": "Votre porte-clés contient plusieurs comptes. Quel compte souhaitez-vous gérer en premier dans l'application ? Vous pourrez en créer ou sélectionner d'autres quand vous le souhaitez.",
 	"activateVibrations": "Activer les vibrations",
 	"activateNotifications": "Activer les notifications",
-	"transactionSignatureCommandReceivedNotification": "L'application %1 souhaite envoyer une transaction.",
+	"transactionSignatureCommandReceivedNotificationTransfer": "L'application %1 souhaite effectuer une action nécessitant votre confirmation.",
 	"transactionInputNotification": "Vous avez reçu %1 %2 sur votre compte %3",
 	"searchField": "Recherche...",
 	"search": "Rechercher",
